--- a/src/Combinatorics/EnumerativeCombinatorics/weak_compositions.jl
+++ b/src/Combinatorics/EnumerativeCombinatorics/weak_compositions.jl
@@ -138,7 +138,7 @@ function Base.iterate(W::WeakCompositions{T}, state::Nothing = nothing) where T
   return weak_composition(c, check = false), s
 end
 
-function Base.iterate(W::WeakCompositions{T}, s::Vector{T}) where T
+@inline function Base.iterate(W::WeakCompositions{T}, s::Vector{T}) where T
   n = base(W)
   k = parts(W)
   if k == 0 || s[k] == n