… docs/src/Combinatorics/EnumerativeCombinatorics/partitions.md
… test/Combinatorics/EnumerativeCombinatorics/partitions.jl