Turns the last (I hope) `partitions` function into a proper iterator.
I split up and tried to simplify the function a bit; the original version is like 200 lines long.

I also added TODOs to the tableaux constructors which need a similar treatment so "we" don't forget about it. (Two of them should be fairly easy to do, but one is a recursive algorithm, so let's see.)