@joschmitt 2024-06-10
I put `@inline` in front of the `iterate` functions. With this, the number of allocations is approximately the number of iterations. Without it, one usually had twice (or even four times) as many allocations.
Here are some timings. It's (almost always) better with `@inline`, sometimes considerably so. Note that some iterators call other iterators internally, so "without inline" might not be without *any* inlining.

### partitions(n)

#### without inline
```
julia> @btime for p in partitions(5) end
  155.104 ns (14 allocations: 960 bytes)

julia> @btime for p in partitions(30) end
  145.177 μs (11208 allocations: 1017.64 KiB)

julia> @btime for p in partitions(50) end
  5.755 ms (408452 allocations: 42.45 MiB)
```

#### with inline
```
julia> @btime for p in partitions(5) end
  126.930 ns (8 allocations: 672 bytes)

julia> @btime for p in partitions(30) end
  113.089 μs (5605 allocations: 755.00 KiB)

julia> @btime for p in partitions(50) end
  4.658 ms (204227 allocations: 33.10 MiB)
```


### partitions(n, k)

#### without inline
```
julia> @btime for p in partitions(5, 3) end
  76.443 ns (6 allocations: 448 bytes)

julia> @btime for p in partitions(30, 10) end
  14.026 μs (1062 allocations: 107.94 KiB)

julia> @btime for p in partitions(100, 50) end
  20.433 ms (408454 allocations: 109.07 MiB)
```

#### with inline
```
julia> @btime for p in partitions(5, 3) end
  66.870 ns (5 allocations: 384 bytes)

julia> @btime for p in partitions(30, 10) end
  11.534 μs (533 allocations: 74.88 KiB)

julia> @btime for p in partitions(100, 50) end
  18.649 ms (204229 allocations: 96.60 MiB)
```


### partitions(n, k, values, multiplicities)

#### without inline
```
julia> @btime for p in partitions(5, 3, collect(1:3), fill(3, 3)) end
  151.944 ns (13 allocations: 784 bytes)

julia> @btime for p in partitions(30, 10, collect(1:10), fill(10, 10)) end
  12.965 μs (1183 allocations: 74.33 KiB)

julia> @btime for p in partitions(100, 50, collect(1:50), fill(50, 50)) end
  21.556 ms (612682 allocations: 105.95 MiB)
```

#### with inline
```
julia> @btime for p in partitions(5, 3, collect(1:3), fill(3, 3)) end
  132.980 ns (9 allocations: 688 bytes)

julia> @btime for p in partitions(30, 10, collect(1:10), fill(10, 10)) end
  9.933 μs (399 allocations: 55.95 KiB)

julia> @btime for p in partitions(100, 50, collect(1:50), fill(50, 50)) end
  19.263 ms (204232 allocations: 96.61 MiB)
```


### partitions(n, values)

#### without inline
```
julia> @btime for p in partitions(5, collect(1:3)) end
  560.263 ns (42 allocations: 2.86 KiB)

julia> @btime for p in partitions(30, collect(1:10)) end
  142.712 μs (10912 allocations: 819.98 KiB)

julia> @btime for p in partitions(50, collect(1:25)) end
  8.292 ms (590911 allocations: 47.27 MiB)
```

#### with inline
```
julia> @btime for p in partitions(5, collect(1:3)) end
  500.634 ns (33 allocations: 2.48 KiB)

julia> @btime for p in partitions(30, collect(1:10)) end
  100.799 μs (3733 allocations: 539.53 KiB)

julia> @btime for p in partitions(50, collect(1:25)) end
  6.000 ms (197136 allocations: 32.25 MiB)
```


### weak_compositions(n, k)

#### without inline
```
julia> @btime for w in weak_compositions(5, 3) end
  389.612 ns (42 allocations: 2.34 KiB)

julia> @btime for w in weak_compositions(10, 5) end
  18.647 μs (2002 allocations: 125.19 KiB)

julia> @btime for w in weak_compositions(20, 10) end
  529.421 ms (20030010 allocations: 1.64 GiB)
```

#### with inline
```
julia> @btime for w in weak_compositions(5, 3) end
  295.093 ns (22 allocations: 1.72 KiB)

julia> @btime for w in weak_compositions(10, 5) end
  14.052 μs (1002 allocations: 93.94 KiB)

julia> @btime for w in weak_compositions(20, 10) end
  448.425 ms (10015006 allocations: 1.34 GiB)
```


### compositions(n, k)

#### without inline
```
julia> @btime for c in compositions(5, 3) end
  125.694 ns (14 allocations: 784 bytes)

julia> @btime for c in compositions(10, 5) end
  2.395 μs (254 allocations: 15.88 KiB)

julia> @btime for c in compositions(20, 10) end
  2.176 ms (184758 allocations: 15.51 MiB)
```

#### with inline
```
julia> @btime for c in compositions(5, 3) end
  97.109 ns (7 allocations: 560 bytes)

julia> @btime for c in compositions(10, 5) end
  1.838 μs (127 allocations: 11.91 KiB)

julia> @btime for c in compositions(20, 10) end
  1.653 ms (92379 allocations: 12.69 MiB)
```

### compositions(n)

#### without inline
```
julia> @btime for c in compositions(5) end
  425.683 ns (41 allocations: 2.81 KiB)

julia> @btime for c in compositions(10) end
  10.796 μs (1043 allocations: 85.27 KiB)

julia> @btime for c in compositions(20) end
  16.656 ms (1048615 allocations: 104.00 MiB)
```

#### with inline
```
julia> @btime for c in compositions(5) end
  348.469 ns (30 allocations: 2.00 KiB)

julia> @btime for c in compositions(10) end
  8.502 μs (541 allocations: 53.61 KiB)

julia> @btime for c in compositions(20) end
  19.439 ms (524347 allocations: 72.00 MiB)
```

It's weird that this is actually slower for `n = 20`.
For `n = 18, 19, 21, 22` the inlined version is faster, so I'll just ignore this outlier.


### ascending_compositions(n)

#### without inline
```
julia> @btime for c in ascending_compositions(5) end
  161.388 ns (16 allocations: 944 bytes)

julia> @btime for c in ascending_compositions(30) end
  147.109 μs (11210 allocations: 930.17 KiB)

julia> @btime for c in ascending_compositions(50) end
  5.832 ms (408454 allocations: 39.34 MiB)
```

#### with inline
```
julia> @btime for c in ascending_compositions(5) end
  132.193 ns (9 allocations: 720 bytes)

julia> @btime for c in ascending_compositions(30) end
  121.606 μs (5606 allocations: 755.05 KiB)

julia> @btime for c in ascending_compositions(50) end
  4.917 ms (204228 allocations: 33.10 MiB)
```