@joschmitt 2024-06-11
I was told anecdotally that someone is preparing a notebook with timings possibly including the partitions code. This pull request might be relevant then, especially the `inline` business (if the notebook is based on `master`).